Refactor the test structure, use `AIAgentTestBase`, place tools to the utils
