+++ integration-tests/src/jvmTest/kotlin/ai/koog/integration/tests/agent/AIAgentMultipleLLMIntegrationTest.kt
@@ -1,62 +1,30 @@
 package ai.koog.integration.tests.agent
 
 import ai.koog.agents.core.agent.AIAgent
-import ai.koog.agents.core.agent.config.AIAgentConfig
-import ai.koog.agents.core.agent.context.agentInput
 import ai.koog.agents.core.agent.exception.AIAgentException
-import ai.koog.agents.core.dsl.builder.forwardTo
-import ai.koog.agents.core.dsl.builder.strategy
-import ai.koog.agents.core.dsl.extension.nodeExecuteTool
-import ai.koog.agents.core.dsl.extension.nodeLLMCompressHistory
-import ai.koog.agents.core.dsl.extension.nodeLLMRequest
-import ai.koog.agents.core.dsl.extension.nodeLLMSendToolResult
-import ai.koog.agents.core.dsl.extension.onAssistantMessage
-import ai.koog.agents.core.dsl.extension.onToolCall
-import ai.koog.agents.core.tools.Tool
-import ai.koog.agents.core.tools.ToolDescriptor
 import ai.koog.agents.core.tools.ToolRegistry
-import ai.koog.agents.core.tools.annotations.LLMDescription
-import ai.koog.agents.ext.agent.subgraphWithTask
 import ai.koog.agents.features.eventHandler.feature.EventHandler
 import ai.koog.agents.features.eventHandler.feature.EventHandlerConfig
-import ai.koog.integration.tests.agent.ReportingLLMLLMClient.Event
+import ai.koog.integration.tests.agent.AIAgentTestBase.ReportingLLMClient.Event
 import ai.koog.integration.tests.utils.Models
 import ai.koog.integration.tests.utils.RetryUtils.withRetry
+import ai.koog.integration.tests.utils.TestUtils.CalculatorTool
+import ai.koog.integration.tests.utils.TestUtils.MockFileSystem
+import ai.koog.integration.tests.utils.TestUtils.OperationResult
 import ai.koog.integration.tests.utils.TestUtils.readTestAnthropicKeyFromEnv
 import ai.koog.integration.tests.utils.TestUtils.readTestOpenAIKeyFromEnv
 import ai.koog.integration.tests.utils.annotations.Retry
-import ai.koog.prompt.dsl.ModerationResult
-import ai.koog.prompt.dsl.Prompt
 import ai.koog.prompt.dsl.prompt
-import ai.koog.prompt.executor.clients.LLMClient
 import ai.koog.prompt.executor.clients.anthropic.AnthropicLLMClient
 import ai.koog.prompt.executor.clients.anthropic.AnthropicModels
 import ai.koog.prompt.executor.clients.openai.OpenAILLMClient
-import ai.koog.prompt.executor.clients.openai.OpenAIModels
 import ai.koog.prompt.executor.llms.MultiLLMPromptExecutor
 import ai.koog.prompt.executor.llms.all.simpleAnthropicExecutor
 import ai.koog.prompt.llm.LLMProvider
 import ai.koog.prompt.llm.LLModel
 import ai.koog.prompt.markdown.markdown
-import ai.koog.prompt.message.Message
-import ai.koog.prompt.params.LLMParams
-import ai.koog.prompt.streaming.StreamFrame
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.DelicateCoroutinesApi
-import kotlinx.coroutines.cancel
 import kotlinx.coroutines.channels.Channel
-import kotlinx.coroutines.coroutineScope
-import kotlinx.coroutines.currentCoroutineContext
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.flow
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.test.TestScope
 import kotlinx.coroutines.test.runTest
-import kotlinx.serialization.KSerializer
-import kotlinx.serialization.Serializable
-import kotlinx.serialization.builtins.serializer
-import org.junit.jupiter.api.AfterEach
-import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.Disabled
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
@@ -74,455 +42,17 @@ import kotlin.test.assertTrue
 import kotlin.test.fail
 import kotlin.time.Duration.Companion.minutes
 
-internal class ReportingLLMLLMClient(
-    private val eventsChannel: Channel<Event>,
-    private val underlyingClient: LLMClient
-) : LLMClient {
-
-    override fun llmProvider(): LLMProvider = underlyingClient.llmProvider()
-    sealed interface Event {
-        data class Message(
-            val llmClient: String,
-            val method: String,
-            val prompt: Prompt,
-            val tools: List<String>,
-            val model: LLModel
-        ) : Event
-
-        data object Termination : Event
-    }
-
-    override suspend fun execute(
-        prompt: Prompt,
-        model: LLModel,
-        tools: List<ToolDescriptor>
-    ): List<Message.Response> {
-        CoroutineScope(currentCoroutineContext()).launch {
-            eventsChannel.send(
-                Event.Message(
-                    llmClient = underlyingClient::class.simpleName ?: "null",
-                    method = "execute",
-                    prompt = prompt,
-                    tools = tools.map { it.name },
-                    model = model
-                )
-            )
-        }
-        return underlyingClient.execute(prompt, model, tools)
-    }
-
-    override fun executeStreaming(
-        prompt: Prompt,
-        model: LLModel,
-        tools: List<ToolDescriptor>
-    ): Flow<StreamFrame> = flow {
-        coroutineScope {
-            eventsChannel.send(
-                Event.Message(
-                    llmClient = underlyingClient::class.simpleName ?: "null",
-                    method = "execute",
-                    prompt = prompt,
-                    tools = emptyList(),
-                    model = model
-                )
-            )
-        }
-        underlyingClient.executeStreaming(prompt, model, tools)
-            .collect(this)
-    }
-
-    override suspend fun moderate(
-        prompt: Prompt,
-        model: LLModel
-    ): ModerationResult {
-        throw NotImplementedError("Moderation not needed for this test")
-    }
-
-    override fun close() {
-        underlyingClient.close()
-        eventsChannel.close()
-    }
-}
-
-internal fun LLMClient.reportingTo(
-    eventsChannel: Channel<Event>
-) = ReportingLLMLLMClient(eventsChannel, this)
-
-class AIAgentMultipleLLMIntegrationTest {
-    private val testScope = TestScope()
-
-    @AfterEach
-    fun cleanup() {
-        testScope.cancel()
-    }
-
+class AIAgentMultipleLLMIntegrationTest : AIAgentTestBase() {
     companion object {
-        private lateinit var testResourcesDir: File
+        @JvmStatic
+        fun getLatestModels(): Stream<LLModel> = AIAgentTestBase.getLatestModels()
 
         @JvmStatic
-        fun getModels(): Stream<LLModel> = Stream.of(
-            AnthropicModels.Sonnet_4_5,
-            OpenAIModels.Chat.GPT5,
-        )
-
-        @JvmStatic
-        @BeforeAll
-        fun setup() {
-            testResourcesDir = File("src/jvmTest/resources/media")
-            testResourcesDir.mkdirs()
-            assertTrue(testResourcesDir.exists(), "Test resources directory should exist")
-        }
-
-        @JvmStatic
-        fun modelsWithVisionCapability(): Stream<Arguments> {
-            return Models.modelsWithVisionCapability()
-        }
+        fun modelsWithVisionCapability(): Stream<Arguments> = AIAgentTestBase.modelsWithVisionCapability()
     }
 
-    // API keys for testing
     private val openAIApiKey: String get() = readTestOpenAIKeyFromEnv()
     private val anthropicApiKey: String get() = readTestAnthropicKeyFromEnv()
-
-    @Serializable
-    enum class CalculatorOperation {
-        ADD,
-        SUBTRACT,
-        MULTIPLY,
-        DIVIDE
-    }
-
-    object CalculatorTool : Tool<CalculatorTool.Args, Int>() {
-        @Serializable
-        data class Args(
-            @property:LLMDescription("The operation to perform.")
-            val operation: CalculatorOperation,
-            @property:LLMDescription("The first argument (number)")
-            val a: Int,
-            @property:LLMDescription("The second argument (number)")
-            val b: Int
-        )
-
-        override val argsSerializer = Args.serializer()
-        override val resultSerializer: KSerializer<Int> = Int.serializer()
-
-        override val name: String = "calculator"
-        override val description: String =
-            "A simple calculator that can add, subtract, multiply, and divide two numbers."
-
-        override suspend fun execute(args: Args): Int = when (args.operation) {
-            CalculatorOperation.ADD -> args.a + args.b
-            CalculatorOperation.SUBTRACT -> args.a - args.b
-            CalculatorOperation.MULTIPLY -> args.a * args.b
-            CalculatorOperation.DIVIDE -> args.a / args.b
-        }
-    }
-
-    sealed interface OperationResult<T> {
-        class Success<T>(val result: T) : OperationResult<T>
-        class Failure<T>(val error: String) : OperationResult<T>
-    }
-
-    class MockFileSystem {
-        private val fileContents: MutableMap<String, String> = mutableMapOf()
-
-        fun create(path: String, content: String): OperationResult<Unit> {
-            if (path in fileContents) return OperationResult.Failure("File already exists")
-            fileContents[path] = content
-            return OperationResult.Success(Unit)
-        }
-
-        fun delete(path: String): OperationResult<Unit> {
-            if (path !in fileContents) return OperationResult.Failure("File does not exist")
-            fileContents.remove(path)
-            return OperationResult.Success(Unit)
-        }
-
-        fun read(path: String): OperationResult<String> {
-            if (path !in fileContents) return OperationResult.Failure("File does not exist")
-            return OperationResult.Success(fileContents[path]!!)
-        }
-
-        fun ls(path: String): OperationResult<List<String>> {
-            if (path in fileContents) {
-                return OperationResult.Failure("Path $path points to a file, but not a directory!")
-            }
-            val matchingFiles = fileContents
-                .filter { (filePath, _) -> filePath.startsWith(path) }
-                .map { (filePath, _) -> filePath }
-
-            if (matchingFiles.isEmpty()) {
-                return OperationResult.Failure("No files in the directory. Directory doesn't exist or is empty.")
-            }
-            return OperationResult.Success(matchingFiles)
-        }
-
-        fun fileCount(): Int = fileContents.size
-    }
-
-    class CreateFile(private val fs: MockFileSystem) : Tool<CreateFile.Args, CreateFile.Result>() {
-        @Serializable
-        data class Args(
-            @property:LLMDescription("The path to create the file")
-            val path: String,
-            @property:LLMDescription("The content to create the file")
-            val content: String
-        )
-
-        @Serializable
-        data class Result(val successful: Boolean, val message: String? = null)
-
-        override val argsSerializer = Args.serializer()
-        override val resultSerializer: KSerializer<Result> = Result.serializer()
-
-        override val name: String = "create_file"
-        override val description: String =
-            "Create a file and writes the given text content to it"
-
-        override suspend fun execute(args: Args): Result {
-            return when (val res = fs.create(args.path, args.content)) {
-                is OperationResult.Success -> Result(successful = true)
-                is OperationResult.Failure -> Result(successful = false, message = res.error)
-            }
-        }
-    }
-
-    class DeleteFile(private val fs: MockFileSystem) : Tool<DeleteFile.Args, DeleteFile.Result>() {
-        @Serializable
-        data class Args(
-            @property:LLMDescription("The path of the file to be deleted")
-            val path: String
-        )
-
-        @Serializable
-        data class Result(val successful: Boolean, val message: String? = null)
-
-        override val argsSerializer = Args.serializer()
-        override val resultSerializer: KSerializer<Result> = Result.serializer()
-
-        override val name: String = "delete_file"
-        override val description: String = "Deletes a file"
-
-        override suspend fun execute(args: Args): Result {
-            return when (val res = fs.delete(args.path)) {
-                is OperationResult.Success -> Result(successful = true)
-                is OperationResult.Failure -> Result(successful = false, message = res.error)
-            }
-        }
-    }
-
-    class ReadFile(private val fs: MockFileSystem) : Tool<ReadFile.Args, ReadFile.Result>() {
-        @Serializable
-        data class Args(
-            @property:LLMDescription("The path of the file to read")
-            val path: String
-        )
-
-        @Serializable
-        data class Result(
-            val successful: Boolean,
-            val message: String? = null,
-            val content: String? = null
-        )
-
-        override val argsSerializer = Args.serializer()
-        override val resultSerializer: KSerializer<Result> = Result.serializer()
-
-        override val name: String = "read_file"
-        override val description: String = "Reads a file"
-
-        override suspend fun execute(args: Args): Result {
-            return when (val res = fs.read(args.path)) {
-                is OperationResult.Success<String> -> Result(successful = true, content = res.result)
-                is OperationResult.Failure -> Result(successful = false, message = res.error)
-            }
-        }
-    }
-
-    class ListFiles(private val fs: MockFileSystem) : Tool<ListFiles.Args, ListFiles.Result>() {
-        @Serializable
-        data class Args(
-            @property:LLMDescription("The path of the directory")
-            val path: String
-        )
-
-        @Serializable
-        data class Result(
-            val successful: Boolean,
-            val message: String? = null,
-            val children: List<String>? = null
-        )
-
-        override val argsSerializer = Args.serializer()
-        override val resultSerializer: KSerializer<Result> = Result.serializer()
-
-        override val name: String = "list_files"
-        override val description: String = "List all files inside the given path of the directory"
-
-        override suspend fun execute(args: Args): Result {
-            return when (val res = fs.ls(args.path)) {
-                is OperationResult.Success<List<String>> -> Result(successful = true, children = res.result)
-                is OperationResult.Failure -> Result(successful = false, message = res.error)
-            }
-        }
-    }
-
-    @OptIn(DelicateCoroutinesApi::class)
-    private fun createTestMultiLLMAgent(
-        fs: MockFileSystem,
-        eventHandlerConfig: EventHandlerConfig.() -> Unit,
-        maxAgentIterations: Int,
-        prompt: Prompt = prompt("test") {},
-        initialExecutor: MultiLLMPromptExecutor? = null,
-    ): AIAgent<String, String> {
-        val executor = if (initialExecutor == null) {
-            val openAIClient = OpenAILLMClient(openAIApiKey)
-            val anthropicClient = AnthropicLLMClient(anthropicApiKey)
-            MultiLLMPromptExecutor(
-                LLMProvider.OpenAI to openAIClient,
-                LLMProvider.Anthropic to anthropicClient
-            )
-        } else {
-            initialExecutor
-        }
-        val strategy = strategy<String, String>("test") {
-            val anthropicSubgraph by subgraph<String, Unit>("anthropic") {
-                val definePromptAnthropic by node<Unit, Unit> {
-                    llm.writeSession {
-                        model = AnthropicModels.Haiku_4_5
-                        rewritePrompt {
-                            prompt("test") {
-                                system {
-                                    +"You are a helpful assistant. You need to solve my task. "
-                                    +"JUST CALL TOOLS. NO QUESTIONS ASKED."
-                                }
-                            }
-                        }
-                    }
-                }
-
-                val callLLM by nodeLLMRequest(allowToolCalls = true)
-                val callTool by nodeExecuteTool()
-                val sendToolResult by nodeLLMSendToolResult()
-
-                edge(nodeStart forwardTo definePromptAnthropic transformed {})
-                edge(definePromptAnthropic forwardTo callLLM transformed { agentInput<String>() })
-                edge(callLLM forwardTo callTool onToolCall { true })
-                edge(callLLM forwardTo nodeFinish onAssistantMessage { true } transformed {})
-                edge(callTool forwardTo sendToolResult)
-                edge(sendToolResult forwardTo callTool onToolCall { true })
-                edge(sendToolResult forwardTo nodeFinish onAssistantMessage { true } transformed {})
-            }
-
-            val openaiSubgraph by subgraph("openai") {
-                val definePromptOpenAI by node<Unit, Unit> {
-                    llm.writeSession {
-                        model = OpenAIModels.Chat.GPT5
-                        rewritePrompt {
-                            prompt("test") {
-                                system(
-                                    """
-                                    You are a helpful assistant. You need to verify that the task is solved correctly.
-                                    Please analyze the whole produced solution, and check that it is valid.
-                                    Write concise verification result.
-                                    JUST CALL TOOLS. NO QUESTIONS ASKED.
-                                    """.trimIndent()
-                                )
-                            }
-                        }
-                    }
-                }
-
-                val callLLM by nodeLLMRequest(allowToolCalls = true)
-                val callTool by nodeExecuteTool()
-                val sendToolResult by nodeLLMSendToolResult()
-
-                edge(nodeStart forwardTo definePromptOpenAI)
-                edge(definePromptOpenAI forwardTo callLLM transformed { agentInput<String>() })
-                edge(callLLM forwardTo callTool onToolCall { true })
-                edge(callLLM forwardTo nodeFinish onAssistantMessage { true })
-                edge(callTool forwardTo sendToolResult)
-                edge(sendToolResult forwardTo callTool onToolCall { true })
-                edge(sendToolResult forwardTo nodeFinish onAssistantMessage { true })
-            }
-
-            val compressHistoryNode by nodeLLMCompressHistory<Unit>("compress_history")
-
-            nodeStart then anthropicSubgraph then compressHistoryNode then openaiSubgraph then nodeFinish
-        }
-
-        val tools = ToolRegistry {
-            tool(CreateFile(fs))
-            tool(DeleteFile(fs))
-            tool(ReadFile(fs))
-            tool(ListFiles(fs))
-        }
-
-        return AIAgent(
-            promptExecutor = executor,
-            strategy = strategy,
-            agentConfig = AIAgentConfig(prompt, OpenAIModels.Chat.GPT5, maxAgentIterations),
-            toolRegistry = tools,
-        ) {
-            install(EventHandler, eventHandlerConfig)
-        }
-    }
-
-    private fun createTestAgentWithToolsInSubgraph(
-        fs: MockFileSystem,
-        eventHandlerConfig: EventHandlerConfig.() -> Unit = {},
-        model: LLModel,
-        emptyAgentRegistry: Boolean = true,
-    ): AIAgent<String, String> {
-        val openAIClient = OpenAILLMClient(openAIApiKey)
-        val anthropicClient = AnthropicLLMClient(anthropicApiKey)
-
-        val executor = MultiLLMPromptExecutor(
-            LLMProvider.OpenAI to openAIClient,
-            LLMProvider.Anthropic to anthropicClient
-        )
-
-        val subgraphTools = listOf(
-            CreateFile(fs),
-            ReadFile(fs),
-            ListFiles(fs),
-            DeleteFile(fs),
-        )
-
-        val strategy = strategy<String, String>("test-subgraph-only-tools") {
-            val fileOperationsSubgraph by subgraphWithTask<String, String>(
-                tools = subgraphTools,
-                llmModel = model,
-                llmParams = LLMParams(toolChoice = LLMParams.ToolChoice.Required)
-            ) { input ->
-                "You are a helpful assistant that can perform file operations. Use the available tools to complete the following task: $input. Make sure to use tools when needed and provide clear feedback about what you've done."
-            }
-
-            nodeStart then fileOperationsSubgraph then nodeFinish
-        }
-
-        val toolRegistry = if (emptyAgentRegistry) {
-            ToolRegistry {}
-        } else {
-            ToolRegistry {
-                subgraphTools.forEach { tool(it) }
-            }
-        }
-
-        return AIAgent(
-            promptExecutor = executor,
-            strategy = strategy,
-            agentConfig = AIAgentConfig(
-                prompt = prompt("test") {
-                    system("You are a helpful assistant.")
-                },
-                model,
-                maxAgentIterations = 50,
-            ),
-            toolRegistry = toolRegistry,
-        ) {
-            install(EventHandler, eventHandlerConfig)
-        }
-    }
 
     @Test
     @Retry(5)
@@ -599,7 +129,7 @@ class AIAgentMultipleLLMIntegrationTest {
     }
 
     @ParameterizedTest
-    @MethodSource("getModels")
+    @MethodSource("getLatestModels")
     @Disabled("See KG-520 Agent with an empty tool registry is stuck into a loop if a subgraph has tools")
     fun `integration_test agent with not registered subgraph tool result fails`(model: LLModel) =
         runTest(timeout = 10.minutes) {
@@ -620,7 +150,7 @@ class AIAgentMultipleLLMIntegrationTest {
         }
 
     @ParameterizedTest
-    @MethodSource("getModels")
+    @MethodSource("getLatestModels")
     fun `integration_test agent with registered subgraph tool result runs`(model: LLModel) =
         runTest(timeout = 10.minutes) {
             Models.assumeAvailable(LLMProvider.OpenAI)
@@ -740,7 +270,7 @@ class AIAgentMultipleLLMIntegrationTest {
         Models.assumeAvailable(model.provider)
         val fs = MockFileSystem()
 
-        val imageFile = File(testResourcesDir, "test.png")
+        val imageFile = File(testResourcesDir.toFile(), "test.png")
         assertTrue(imageFile.exists(), "Image test file should exist")
 
         val imageBytes = imageFile.readBytes()
@@ -799,7 +329,7 @@ class AIAgentMultipleLLMIntegrationTest {
             }
         }
 
-        val imageFile = File(testResourcesDir, "test.png")
+        val imageFile = File(testResourcesDir.toFile(), "test.png")
         assertTrue(imageFile.exists(), "Image test file should exist")
 
         val prompt = prompt("example-prompt") {
