+++ integration-tests/src/jvmTest/kotlin/ai/koog/integration/tests/capabilities/ModelCapabilitiesIntegrationTest.kt
@@ -7,7 +7,7 @@ import ai.koog.integration.tests.utils.MediaTestUtils.createVideoFileForScenario
 import ai.koog.integration.tests.utils.MediaTestUtils.getImageFileForScenario
 import ai.koog.integration.tests.utils.Models
 import ai.koog.integration.tests.utils.RetryUtils.withRetry
-import ai.koog.integration.tests.utils.TestUtils.CalculatorTool
+import ai.koog.integration.tests.utils.TestUtils.SimpleCalculatorTool
 import ai.koog.integration.tests.utils.TestUtils.assertExceptionMessageContains
 import ai.koog.integration.tests.utils.TestUtils.isValidJson
 import ai.koog.integration.tests.utils.TestUtils.readTestAnthropicKeyFromEnv
@@ -140,7 +140,7 @@ class ModelCapabilitiesIntegrationTest {
                 }
 
                 LLMCapability.Tools, LLMCapability.ToolChoice -> {
-                    val tools = CalculatorTool.descriptor
+                    val tools = SimpleCalculatorTool.descriptor
                     val prompt = prompt("cap-tools-positive", params = LLMParams(toolChoice = ToolChoice.Required)) {
                         system("You are a helpful assistant with a calculator tool. Always use the tool.")
                         user("Compute 2 + 3.")
@@ -373,7 +373,7 @@ class ModelCapabilitiesIntegrationTest {
                 }
 
                 LLMCapability.Tools, LLMCapability.ToolChoice -> {
-                    val tools = CalculatorTool.descriptor
+                    val tools = SimpleCalculatorTool.descriptor
                     val prompt = prompt("cap-tools-negative", params = LLMParams(toolChoice = ToolChoice.Required)) {
                         system("You are a helpful assistant with a calculator tool. Always use the tool.")
                         user("Compute 2 + 3.")
